Proper handling of output to stdout
in Python 3 and on Windows. Stdout in text output (as opposed to binary).
It expects a string (not bytes) in Python 3. In Python 2 on Windows, any '\x0a' byte
will be converted  by '\x0d\x0a' when using stdout text output. This can be suppressed
by using the -u command line option.